+++ src/organizer/qorganizeritemsortorder.cpp
@@ -138,7 +138,7 @@ QDataStream &operator<<(QDataStream &out, const QOrganizerItemSortOrder &sortOrd
 {
     quint8 formatVersion = 1; // Version of QDataStream format for QOrganizerItemSortOrder
     return out << formatVersion
-               << sortOrder.detailType()
+               << static_cast<quint32>(sortOrder.detailType())
                << sortOrder.detailField()
                << static_cast<quint32>(sortOrder.blankPolicy())
                << static_cast<quint32>(sortOrder.direction())
